refactor(core)(!): reworked message response API to make it more type-safe
BREAKING CHANGE: This will break any code that was relying on the existence of the Response class.
This also currently causes a bunch of spurious rewrite test failures, because some FNGs are strange and contain semantically invalid message responses.
The invalid parts are ignored, causing slight differences between the original and rewritten FNGs. Whether this is actually an issue remains to be seen.